requirements.txt
# Core parsers
opendataloader-pdf>=0.5.0
opendataloader-pdf[hybrid]>=0.5.0
docling>=2.0.0
marker-pdf>=1.0.0
pymupdf4llm>=0.0.17
unstructured[pdf]>=0.14.0
unstructured[local-inference]>=0.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65

# Evaluation
pandas>=2.0
rapidfuzz>=3.0
jinja2>=3.1
tabulate>=0.9
